replaces poseidon with circomposeidon
