+++ arkworks-gadgets/src/poseidon/constraints.rs
@@ -1,22 +1,17 @@
-use super::CRH;
+use crate::poseidon::CRH;
 use ark_crypto_primitives::crh::constraints::{CRHGadget as CRHGadgetTrait, TwoToOneCRHGadget};
 use ark_ff::PrimeField;
 use ark_r1cs_std::{
-	alloc::AllocVar,
 	fields::{fp::FpVar, FieldVar},
-	prelude::*,
 	uint8::UInt8,
 };
-use ark_relations::r1cs::{Namespace, SynthesisError};
-use ark_std::{marker::PhantomData, vec::Vec};
-use arkworks_utils::{
-	poseidon::{
-		sbox::{constraints::SboxConstraints, PoseidonSbox},
-		PoseidonParameters,
-	},
-	utils::to_field_var_elements,
-};
+use ark_r1cs_std::{alloc::AllocVar, prelude::*};
 use core::borrow::Borrow;
+use ark_relations::r1cs::SynthesisError;
+use ark_relations::r1cs::Namespace;
+use ark_std::{marker::PhantomData, vec::Vec};
+use arkworks_utils::poseidon::{PoseidonParameters, sbox::{PoseidonSbox, constraints::SboxConstraints}};
+use core::ops::{Add, AddAssign, Mul};
 
 #[derive(Default, Clone)]
 pub struct PoseidonParametersVar<F: PrimeField> {
@@ -32,125 +27,6 @@ pub struct PoseidonParametersVar<F: PrimeField> {
 	pub width: u8,
 	/// The S-box to apply in the sub words layer.
 	pub sbox: PoseidonSbox,
-}
-
-pub struct CRHGadget<F: PrimeField> {
-	field: PhantomData<F>,
-}
-
-impl<F: PrimeField> CRHGadget<F> {
-	fn permute(
-		parameters: &PoseidonParametersVar<F>,
-		mut state: Vec<FpVar<F>>,
-	) -> Result<Vec<FpVar<F>>, SynthesisError> {
-		let width = parameters.width as usize;
-
-		let mut round_keys_offset = 0;
-
-		// full Sbox rounds
-		for _ in 0..(parameters.full_rounds / 2) {
-			// Substitution (S-box) layer
-			for i in 0..width {
-				state[i] += &parameters.round_keys[round_keys_offset];
-				state[i] = parameters.sbox.synthesize_sbox(&state[i])?;
-				round_keys_offset += 1;
-			}
-			// Apply linear layer
-			state = Self::apply_linear_layer(&state, &parameters.mds_matrix);
-		}
-
-		// middle partial Sbox rounds
-		for _ in 0..parameters.partial_rounds {
-			// Substitution (S-box) layer
-			for i in 0..width {
-				state[i] += &parameters.round_keys[round_keys_offset];
-				round_keys_offset += 1;
-			}
-			// apply Sbox to only 1 element of the state.
-			// Here the last one is chosen but the choice is arbitrary.
-			state[0] = parameters.sbox.synthesize_sbox(&state[0])?;
-			// Linear layer
-			state = Self::apply_linear_layer(&state, &parameters.mds_matrix);
-		}
-
-		// last full Sbox rounds
-		for _ in 0..(parameters.full_rounds / 2) {
-			// Substitution (S-box) layer
-			for i in 0..width {
-				state[i] += &parameters.round_keys[round_keys_offset];
-				state[i] = parameters.sbox.synthesize_sbox(&state[i])?;
-				round_keys_offset += 1;
-			}
-			// Linear layer
-			state = Self::apply_linear_layer(&state, &parameters.mds_matrix);
-		}
-
-		Ok(state)
-	}
-
-	fn apply_linear_layer(state: &Vec<FpVar<F>>, mds_matrix: &Vec<Vec<FpVar<F>>>) -> Vec<FpVar<F>> {
-		let mut new_state: Vec<FpVar<F>> = Vec::new();
-		for i in 0..state.len() {
-			let mut sc = FpVar::<F>::zero();
-			for j in 0..state.len() {
-				let mij = &mds_matrix[i][j];
-				sc += mij * &state[j];
-			}
-			new_state.push(sc);
-		}
-		new_state
-	}
-}
-
-// https://github.com/arkworks-rs/r1cs-std/blob/master/src/bits/uint8.rs#L343
-impl<F: PrimeField> CRHGadgetTrait<CRH<F>, F> for CRHGadget<F> {
-	type OutputVar = FpVar<F>;
-	type ParametersVar = PoseidonParametersVar<F>;
-
-	fn evaluate(
-		parameters: &Self::ParametersVar,
-		input: &[UInt8<F>],
-	) -> Result<Self::OutputVar, SynthesisError> {
-		let f_var_inputs: Vec<FpVar<F>> = to_field_var_elements(input)?;
-		let width = parameters.width as usize;
-
-		if f_var_inputs.len() > width {
-			panic!(
-				"incorrect input length {:?} for width {:?}",
-				f_var_inputs.len(),
-				parameters.width,
-			);
-		}
-
-		let mut buffer = vec![FpVar::zero(); width];
-		buffer
-			.iter_mut()
-			.zip(f_var_inputs)
-			.for_each(|(b, l_b)| *b = l_b);
-
-		let result = Self::permute(&parameters, buffer);
-		result.map(|x| x.get(0).cloned().ok_or(SynthesisError::AssignmentMissing))?
-	}
-}
-
-impl<F: PrimeField> TwoToOneCRHGadget<CRH<F>, F> for CRHGadget<F> {
-	type OutputVar = FpVar<F>;
-	type ParametersVar = PoseidonParametersVar<F>;
-
-	fn evaluate(
-		parameters: &Self::ParametersVar,
-		left_input: &[UInt8<F>],
-		right_input: &[UInt8<F>],
-	) -> Result<Self::OutputVar, SynthesisError> {
-		// assume equality of left and right length
-		assert_eq!(left_input.len(), right_input.len());
-		let chained_input: Vec<_> = left_input
-			.to_vec()
-			.into_iter()
-			.chain(right_input.to_vec().into_iter())
-			.collect();
-		<Self as CRHGadgetTrait<_, _>>::evaluate(parameters, &chained_input)
-	}
 }
 
 impl<F: PrimeField> AllocVar<PoseidonParameters<F>, F> for PoseidonParametersVar<F> {
@@ -188,25 +64,118 @@ impl<F: PrimeField> AllocVar<PoseidonParameters<F>, F> for PoseidonParametersVar
 	}
 }
 
+pub struct CRHGadget<F: PrimeField>(PhantomData<F>);
+
+impl<F: PrimeField> CRHGadget<F> {
+	fn permute(
+		params: &PoseidonParametersVar<F>,
+		mut state: Vec<FpVar<F>>,
+	) -> Result<Vec<FpVar<F>>, SynthesisError> {
+		let nr = (params.full_rounds + params.partial_rounds) as usize;
+		for r in 0..nr {
+			state.iter_mut().enumerate().for_each(|(i, a)| {
+				let c = &params.round_keys[(r * (params.width as usize) + i)];
+				a.add_assign(c);
+			});
+
+			let half_rounds = (params.full_rounds as usize) / 2;
+			if r < half_rounds || r >= half_rounds + (params.partial_rounds as usize) {
+				state
+					.iter_mut()
+					.try_for_each(|a| params.sbox.synthesize_sbox(a).map(|f| *a = f))?;
+			} else {
+				state[0] = params.sbox.synthesize_sbox(&state[0])?;
+			}
+
+			state = state
+				.iter()
+				.enumerate()
+				.map(|(i, _)| {
+					state
+						.iter()
+						.enumerate()
+						.fold(FpVar::<F>::zero(), |acc, (j, a)| {
+							let m = &params.mds_matrix[i][j];
+							acc.add(m.mul(a))
+						})
+				})
+				.collect();
+		}
+		Ok(state)
+	}
+}
+
+impl<F: PrimeField> CRHGadgetTrait<CRH<F>, F> for CRHGadget<F> {
+	type OutputVar = FpVar<F>;
+	type ParametersVar = PoseidonParametersVar<F>;
+
+	fn evaluate(
+		parameters: &Self::ParametersVar,
+		input: &[UInt8<F>],
+	) -> Result<Self::OutputVar, SynthesisError> {
+		let f_var_inputs = arkworks_utils::utils::to_field_var_elements(input)?;
+		if f_var_inputs.len() >= parameters.width.into() {
+			panic!(
+				"incorrect input length {:?} for width {:?} -- input bits {:?}",
+				f_var_inputs.len(),
+				parameters.width,
+				input.len()
+			);
+		}
+
+		let mut buffer = vec![FpVar::zero()];
+		for f in f_var_inputs {
+			buffer.push(f);
+		}
+		let result = Self::permute(&parameters, buffer);
+
+		result.map(|x| x.get(0).cloned().ok_or(SynthesisError::AssignmentMissing))?
+	}
+}
+
+impl<F: PrimeField> TwoToOneCRHGadget<CRH<F>, F> for CRHGadget<F> {
+	type OutputVar = FpVar<F>;
+	type ParametersVar = PoseidonParametersVar<F>;
+
+	fn evaluate(
+		parameters: &Self::ParametersVar,
+		left_input: &[UInt8<F>],
+		right_input: &[UInt8<F>],
+	) -> Result<Self::OutputVar, SynthesisError> {
+		// assume equality of left and right length
+		assert_eq!(left_input.len(), right_input.len());
+		let chained_input: Vec<_> = left_input
+			.to_vec()
+			.into_iter()
+			.chain(right_input.to_vec().into_iter())
+			.collect();
+		<Self as CRHGadgetTrait<_, _>>::evaluate(parameters, &chained_input)
+	}
+}
+
 #[cfg(test)]
 mod test {
 	use super::*;
 	use ark_crypto_primitives::crh::CRH as CRHTrait;
-	use ark_ed_on_bls12_381::Fq;
-	use ark_ff::{to_bytes, Zero};
+	use ark_ed_on_bn254::Fq;
+	use ark_ff::to_bytes;
+	use ark_r1cs_std::{
+		alloc::{AllocVar, AllocationMode},
+		R1CSVar,
+	};
 	use ark_relations::r1cs::ConstraintSystem;
-	use arkworks_utils::utils::common::{setup_params_x5_3, Curve};
+	use arkworks_utils::utils::common::setup_circom_params_x5_3;
 
 	type PoseidonCRH3 = CRH<Fq>;
 	type PoseidonCRH3Gadget = CRHGadget<Fq>;
 
 	#[test]
-	fn test_poseidon_native_equality() {
+	fn circom_poseidon_native_equality() {
 		let cs = ConstraintSystem::<Fq>::new_ref();
 
-		let curve = Curve::Bls381;
+		let curve = arkworks_utils::utils::common::Curve::Bn254;
 
-		let params = setup_params_x5_3(curve);
+		let params = setup_circom_params_x5_3(curve);
 
 		let params_var = PoseidonParametersVar::new_variable(
 			cs.clone(),
@@ -215,15 +184,13 @@ mod test {
 		)
 		.unwrap();
 
-		// Test Poseidon on an input of 3 field elements. This will not require padding,
-		// since the inputs are aligned to the expected input chunk size of 32.
-		let aligned_inp = to_bytes![Fq::zero(), Fq::from(1u128), Fq::from(2u128)].unwrap();
+		let aligned_inp = to_bytes![Fq::from(1u128), Fq::from(2u128)].unwrap();
 		let aligned_inp_var =
 			Vec::<UInt8<Fq>>::new_input(cs.clone(), || Ok(aligned_inp.clone())).unwrap();
 
 		let res = PoseidonCRH3::evaluate(&params, &aligned_inp).unwrap();
 		let res_var = <PoseidonCRH3Gadget as CRHGadgetTrait<_, _>>::evaluate(
-			&params_var.clone(),
+			&params_var,
 			&aligned_inp_var,
 		)
 		.unwrap();
@@ -233,11 +200,11 @@ mod test {
 		// inputs are not aligned to the expected input chunk size of 32.
 		let unaligned_inp: Vec<u8> = vec![1, 2, 3, 4, 5, 6];
 		let unaligned_inp_var =
-			Vec::<UInt8<Fq>>::new_input(cs.clone(), || Ok(unaligned_inp.clone())).unwrap();
+			Vec::<UInt8<Fq>>::new_input(cs, || Ok(unaligned_inp.clone())).unwrap();
 
 		let res = PoseidonCRH3::evaluate(&params, &unaligned_inp).unwrap();
 		let res_var = <PoseidonCRH3Gadget as CRHGadgetTrait<_, _>>::evaluate(
-			&params_var.clone(),
+			&params_var,
 			&unaligned_inp_var,
 		)
 		.unwrap();
